refactor: update import paths for countryData.jsx
